Small stuff
Rendered in the register. Forgot to ask Nathan for the texture map, so I made a pink texture map(hence why the register is pink).
The duck is still there because I was using it as a reference.
Was trying to create some functions to build meshes and materials but that proved fruitless.